Add security for cud database endpoint

+++ src/main/kotlin/com/example/militaryservicecompanychecker/company/controller/CompanyController.kt
@@ -10,13 +10,15 @@ import com.example.militaryservicecompanychecker.company.enums.ServiceType
 import com.example.militaryservicecompanychecker.company.service.CompanyService
 import com.example.militaryservicecompanychecker.company.util.Util.safeValueOf
 import org.springframework.http.MediaType
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
 import org.springframework.web.bind.annotation.*
 import org.springframework.web.multipart.MultipartFile
 
 @RestController
 @CrossOrigin(origins = ["https://k-agent.services/", "http://localhost:3000/"])
 class CompanyController(
-    private val companyService: CompanyService
+    private val companyService: CompanyService,
+    private val passwordEncoder: BCryptPasswordEncoder
 ) {
     @PostMapping("/search/autocomplete")
     fun searchCompanyByRegex(@RequestBody request: CompanyAutoCompleteRequest): CompanyAutoCompleteResponse {
@@ -59,9 +61,16 @@ class CompanyController(
         ]
     )
     fun updateCompany(
+        @RequestPart("password") password: String,
         @RequestPart("serviceType") serviceType: String,
         @RequestPart("file") file: MultipartFile
-    ): CompanyResponse {
+    ): Any {
+        if (!passwordEncoder.matches(
+                password,
+                System.getenv("ADMIN_PASSWORD")
+            )
+        ) return "비밀번호가 틀렸습니다."
+
         return CompanyResponse(
             companyService.deleteAndCreateCompanyByFile(file, ServiceType[serviceType]!!)
         )
